refactor converter methods
